requirements/base.txt
Django==1.6.5
Markdown==2.4.1
Pygments==1.6
boto==2.31.1
-e git+https://github.com/ottoyiu/django-cors-headers.git
django-storages==1.1.8
django-subdomains==2.0.4
django-widget-tweaks==1.3
djangorestframework==2.3.14
